implent mvc model
