+++ models/url.js
@@ -1,0 +1,11 @@
+const mongoose = require('mongoose');
+const Schema = mongose.Schema;
+
+let urlSchema = new Schema ({
+  id: String,
+  url: String,
+  date: { type: Date, default: Date.now },
+  count: Number
+});
+
+module.exports = mongoose.model('Url', urlSchema);
