+++ package.json
@@ -25,8 +25,5 @@
     "mongoose": "^4.7.0",
     "shortid": "^2.2.6"
   },
-  "devDependencies": {
-    "webpack": "^1.13.3",
-    "webpack-dev-server": "^1.16.2"
-  }
+  "devDependencies": {}
 }
