--- public/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Jet Pack</title>
-  </head>
-  <body>
-    <script src="main.bundle.js"></script>
-  </body>
-</html>
