+++ views/index.ejs
@@ -6,7 +6,7 @@
   </head>
   <body>
     <h1>Jet Pack</h1>
-    <form action="dinosaurs" method="POST">
+    <form action="/jetpack" method="POST">
       <input type='text' placeholder='URL'></input>
       <input type='submit'></input>
     </form>
