--- webpack.config.js
@@ -1,12 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  entry: {
-    main: "./lib/index.js"
-    // test: "mocha!./test/index.js"
-  },
-  output: {
-    path: __dirname,
-    filename: "[name].bundle.js"
-  }
-};
